% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster_example}
\alias{cluster_example}
\title{Cluster example}
\format{An object of class \code{matrix} with 500 rows and 2 columns.}
\usage{
cluster_example
}
\description{
An example, 2-dimensional dataset containing the 100 points for each of five
  bivariate normal distributions arranged equidistant along the outline of a
  circle.
}
\details{
To inspect execute \code{plot(cluster_example)}.
}
\keyword{datasets}
