\name{cstab-package}
\alias{cstab-package}
\alias{cstab}
\docType{package}
\title{
cstab: Selection of Number of Clusters via Normalized Clustering Instability
}
\description{
Selection of the number of clusters in cluster analysis using stability methods
}
\details{
\tabular{ll}{
Package: \tab cstab\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2016-07-26\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jonas Haslbeck

Maintainer: Jonas Haslbeck <jonashaslbeck@gmail.com>
}
