% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstab.R
\docType{package}
\name{cstab}
\alias{cstab}
\alias{cstab-package}
\title{cstab: Selection of number of clusters via normalized clustering instability}
\description{
Selection of the number of clusters in cluster analysis using stability
   methods.
}
\details{
\tabular{ll}{
   Package: \tab cstab\cr
   Type: \tab Package\cr
   Version: \tab 0.01\cr
   Date: \tab 2016-07-26\cr
   License: \tab GPL (>= 2)\cr
   }
}
\author{
Dirk U. Wulff <dirk.wulff@gmail.com>
   Jonas M. B. Haslbeck <jonas.haslbeck@gmail.com>
}

