% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookup}
\alias{lookup}
\title{Create lookup table}
\usage{
lookup(n = 10000L, root = 200)
}
\arguments{
\item{n}{integer specifying the number of}

\item{root}{numeric specifying the root used to avoid
machine limit.}
}
\description{
Create lookup table for faculties
}
