% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cstab.R
\name{plot.cstab}
\alias{plot.cstab}
\title{Plot method for cstab objects}
\usage{
\method{plot}{cstab}(x, ...)
}
\arguments{
\item{x}{a cstab object (output of functions \code{cStability}).}

\item{\dots}{additional arguments passed to print.}
}
\description{
\code{plot.cstab} plots \emph{instability} path.
}
\author{
Jonas M. B. Haslbeck <jonas.haslbeck@gmail.com>
   Dirk U. Wulff <dirk.wulff@gmail.com>
}

