% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cstab.R
\name{print.cstab}
\alias{print.cstab}
\title{Print method for cstab objects}
\usage{
\method{print}{cstab}(x, ...)
}
\arguments{
\item{x}{a cstab object (output of functions \code{cStability}).}

\item{\dots}{additional arguments passed to print.}
}
\description{
\code{print.cstab} prints key variables of cstab objects.
}
\author{
Jonas M. B. Haslbeck <jonas.haslbeck@gmail.com>
   Dirk U. Wulff <dirk.wulff@gmail.com>
}
