\name{print.cstab}
\alias{print.cstab}
\title{
Print method for cstab objects}

\description{
Prints a selection of the output in cstab objects.}

\usage{
\method{print}{cstab}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{cstab object (output of functions \code{cStability}, \code{cDistance}.)}
  \item{\dots}{additional print arguments}
}

\author{
Jonas Haslbeck <jonashaslbeck@gmail.com>}

